requirements.txt
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
distlib==0.3.6
Django==4.1.7
django-filter==23.2
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.9.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
Markdown==3.4.4
mysql-connector-python==8.0.32
numpy==1.24.3
outcome==1.2.0
Pillow==9.5.0
platformdirs==3.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.18.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
trio==0.22.2
trio-websocket==0.10.4
TwoCaptcha==0.0.1
tzdata==2022.7
undetected-chromedriver==3.5.3
urllib3==2.0.4
virtualenv==20.20.0
websockets==11.0.3
wsproto==1.2.0
zipp==3.16.2
